+++ nemo_retriever/src/nemo_retriever/recall/core.py
@@ -17,7 +17,6 @@ from nemo_retriever.retriever import Retriever
 logger = logging.getLogger(__name__)
 AUDIO_MATCH_TOLERANCE_SECS = 2.0
 
-import numpy as np
 import pandas as pd
 
 
@@ -228,77 +227,6 @@ def _normalize_query_df(df: pd.DataFrame, *, match_mode: str) -> pd.DataFrame:
     df["page"] = df["page"].astype(str)
     df["golden_answer"] = df.apply(lambda x: f"{x.pdf}_{x.page}", axis=1)
     return df
-
-
-def _resolve_embedding_endpoint(cfg: RecallConfig) -> Tuple[Optional[str], Optional[bool]]:
-    """
-    Resolve which embedding endpoint to use.
-
-    Returns (endpoint, use_grpc) where:
-      - endpoint is either an http(s) URL or a host:port string for gRPC
-      - use_grpc is True for gRPC, False for HTTP, None when no endpoint is configured
-    """
-    http_ep = (cfg.embedding_http_endpoint or "").strip() if isinstance(cfg.embedding_http_endpoint, str) else None
-    grpc_ep = (cfg.embedding_grpc_endpoint or "").strip() if isinstance(cfg.embedding_grpc_endpoint, str) else None
-    single = (cfg.embedding_endpoint or "").strip() if isinstance(cfg.embedding_endpoint, str) else None
-
-    if http_ep:
-        return http_ep, False
-    if grpc_ep:
-        return grpc_ep, True
-    if single:
-        # Infer protocol: if a URL scheme is present, treat as HTTP; otherwise gRPC.
-        return single, (not single.lower().startswith("http"))
-
-    return None, None
-
-
-def _embed_queries_nim(
-    queries: List[str],
-    *,
-    endpoint: str,
-    model: str,
-    api_key: str,
-    grpc: bool,
-) -> List[List[float]]:
-    from nv_ingest_api.util.nim import infer_microservice
-
-    # `infer_microservice` returns a list of embeddings.
-    embeddings = infer_microservice(
-        queries,
-        model_name=model,
-        embedding_endpoint=endpoint,
-        nvidia_api_key=(api_key or "").strip(),
-        grpc=bool(grpc),
-        input_type="query",
-    )
-    # Some backends return numpy arrays; normalize to list-of-list floats.
-    out: List[List[float]] = []
-    for e in embeddings:
-        if isinstance(e, np.ndarray):
-            out.append(e.astype("float32").tolist())
-        else:
-            out.append(list(e))
-    return out
-
-
-def _embed_queries_local_hf(
-    queries: List[str],
-    *,
-    device: Optional[str],
-    cache_dir: Optional[str],
-    batch_size: int,
-    model_name: Optional[str] = None,
-) -> List[List[float]]:
-    from nemo_retriever.model import create_local_embedder, is_vl_embed_model
-
-    embedder = create_local_embedder(model_name, device=device, hf_cache_dir=cache_dir)
-
-    if is_vl_embed_model(model_name):
-        vecs = embedder.embed_queries(queries, batch_size=int(batch_size))
-    else:
-        vecs = embedder.embed(["query: " + q for q in queries], batch_size=int(batch_size))
-    return vecs.detach().to("cpu").tolist()
 
 
 def _hits_to_keys(raw_hits: List[List[Dict[str, Any]]]) -> List[List[str]]:
@@ -517,12 +445,12 @@ def retrieve_and_score(
 
     queries = df_query["query"].astype(str).tolist()
     gold = df_query["golden_answer"].astype(str).tolist()
-    endpoint, use_grpc = _resolve_embedding_endpoint(cfg)
     retriever = Retriever(
         lancedb_uri=cfg.lancedb_uri,
         lancedb_table=cfg.lancedb_table,
         embedder=cfg.embedding_model or VL_EMBED_MODEL,
-        embedding_http_endpoint=endpoint,
+        embedding_http_endpoint=cfg.embedding_http_endpoint,
+        embedding_endpoint=cfg.embedding_endpoint,
         embedding_api_key=cfg.embedding_api_key,
         top_k=cfg.top_k,
         nprobes=cfg.nprobes,
